fix firefox signin bug where the user had to signin 2x by running signin in background.ts and using messaging from authcontext like with chrome.. laso did npx prettier --write .
