+++ entrypoints/offscreen/index.html
@@ -1,3 +1,2 @@
-<!DOCTYPE html>
+<!doctype html>
 <script type="module" src="./main.ts"></script>
-
