+++ public/signInWithPopup.html
@@ -1,8 +1,10 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
   <head>
     <title>signInWithPopup</title>
     <script type="module" src="signInWithPopup.js"></script>
   </head>
-  <body><h1>signInWithPopup</h1></body>
+  <body>
+    <h1>signInWithPopup</h1>
+  </body>
 </html>
