+++ utils/messaging.ts
@@ -5,6 +5,7 @@ export type ExtUserInfo = Pick<User, "displayName" | "photoURL" | "email">;
 interface MessagingProtocol {
   // UI->Background messages
   "auth:signIn": () => User | null;
+  "auth:signInFirefox": () => Promise<User | null>;
   "auth:signOut": () => Promise<void>;
   "auth:getUser": () => User | null;
   // Background->UI messages
