+++ wxt.config.ts
@@ -21,4 +21,3 @@ export default defineConfig({
     };
   },
 });
-
